# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Real Estate CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── RESALE ──────────────────────────────────────────────────────────────────
  /resale:
    get:
      operationId: listResaleUnits
      tags: [resale]
      summary: List resale units
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: unitType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resale units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResaleUnit"
    post:
      operationId: createResaleUnit
      tags: [resale]
      summary: Create resale unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResaleUnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResaleUnit"
  /resale/{unitId}:
    get:
      operationId: getResaleUnit
      tags: [resale]
      summary: Get resale unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResaleUnit"
    patch:
      operationId: updateResaleUnit
      tags: [resale]
      summary: Update resale unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResaleUnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResaleUnit"
    delete:
      operationId: deleteResaleUnit
      tags: [resale]
      summary: Delete resale unit
      parameters:
        - name: unitId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── CLIENTS ──────────────────────────────────────────────────────────────────
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Sales performance report
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/leads:
    get:
      operationId: getLeadsReport
      tags: [reports]
      summary: Lead source & conversion report
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsReport"
components:
  schemas:
...
    Project:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        ownerName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        avgPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        leadsCount:
          type: integer
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        ownerName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        avgPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        ownerName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        avgPrice:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── RESALE ────────────────────────────────────────────────────────────────
    ResaleUnit:
      type: object
      required: [id, projectName, isActive, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: ["string", "null"]
        projectName:
          type: string
        area:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        unitType:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    ResaleUnitInput:
      type: object
      required: [projectName]
      properties:
        projectId:
          type: ["string", "null"]
        projectName:
          type: string
          minLength: 1
        area:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        unitType:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
    ResaleUnitUpdate:
      type: object
      properties:
        projectId:
          type: ["string", "null"]
        projectName:
          type: string
        area:
          type: ["string", "null"]
        price:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        unitType:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        ownerPhone:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── CLIENT ────────────────────────────────────────────────────────────────
...
    SalesReport:
      type: object
      required: [byUser, totalWon, totalLost, totalLeads]
      properties:
        totalWon:
          type: integer
        totalLost:
          type: integer
        totalLeads:
          type: integer
        byUser:
          type: array
          items:
            $ref: "#/components/schemas/UserSalesStats"
    UserSalesStats:
      type: object
      required: [userId, userName, won, lost, inProgress]
      properties:
        userId:
          type: string
        userName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        won:
          type: integer
        lost:
          type: integer
        inProgress:
          type: integer
        total:
          type: integer
    LeadsReport:
      type: object
      required: [bySource, byStatus, total]
      properties:
        total:
          type: integer
        bySource:
          type: array
          items:
            $ref: "#/components/schemas/SourceStat"
        byStatus:
          type: array
          items:
            $ref: "#/components/schemas/StatusStat"
    SourceStat:
      type: object
      required: [source, count]
      properties:
        source:
          type: string
        count:
          type: integer
    StatusStat:
      type: object
      required: [status, count]
      properties:
        status:
          type: string
        count:
          type: integer